requirements.txt
skafossdk>=1.1.7
http://download.pytorch.org/whl/cpu/torch-0.4.1-cp36-cp36m-linux_x86_64.whl
torchvision
neo4j-driver
pandas>=0.23.1
numpy
annoy
scipy
s3fs
